Fixes for PR (remove commented code and moved styles)
